AddressElement and AddressList with new component structure
